int(ceil -> ceil

+++ daal4py/sklearn/ensemble/_forest.py
@@ -814,15 +814,15 @@ class RandomForestClassifier(RandomForestClassifier_original, RandomForestBase):
         if isinstance(self.min_samples_split, numbers.Integral):
             parameters["minObservationsInSplitNode"] = self.min_samples_split
         else:
-            parameters["minObservationsInSplitNode"] = int(
-                ceil(self.min_samples_split * X.shape[0])
+            parameters["minObservationsInSplitNode"] = ceil(
+                self.min_samples_split * X.shape[0]
             )
 
         if isinstance(self.min_samples_leaf, numbers.Integral):
             parameters["minObservationsInLeafNode"] = self.min_samples_leaf
         else:
-            parameters["minObservationsInLeafNode"] = int(
-                ceil(self.min_samples_leaf * X.shape[0])
+            parameters["minObservationsInLeafNode"] = ceil(
+                self.min_samples_leaf * X.shape[0]
             )
 
         if self.bootstrap:
@@ -1350,15 +1350,15 @@ class RandomForestRegressor(RandomForestRegressor_original, RandomForestBase):
         if isinstance(self.min_samples_split, numbers.Integral):
             parameters["minObservationsInSplitNode"] = self.min_samples_split
         else:
-            parameters["minObservationsInSplitNode"] = int(
-                ceil(self.min_samples_split * X.shape[0])
+            parameters["minObservationsInSplitNode"] = ceil(
+                self.min_samples_split * X.shape[0]
             )
 
         if isinstance(self.min_samples_leaf, numbers.Integral):
             parameters["minObservationsInLeafNode"] = self.min_samples_leaf
         else:
-            parameters["minObservationsInLeafNode"] = int(
-                ceil(self.min_samples_leaf * X.shape[0])
+            parameters["minObservationsInLeafNode"] = ceil(
+                self.min_samples_leaf * X.shape[0]
             )
 
         if self.bootstrap:
